add negative condition (!T in Type1, Type2) or (!T in U)
This required a major modification and the use of a real attribute for trait_gen_if. It's scanned
a first time in the #[trait_gen], where the arguments are replaced by types, and checked in the
attribute itself to see if the argument on the left is in the list on the right.